[SYSTEMDS-3451] Fix missing NNZ propagation in federated instructions
For many federated operations that output federated data, no
information on the number of non-zeros was propagated back to the
coordinator and hence, suboptional plan choices where made during
dynamic recompilation. We now generally return the nnz of instruction
outputs from all EXEC_INST and EXEC_UDF types, and provide utils to
obtain this info from the federated responses. The exploitation of
this info was already introduced into right indexing, append, and
replace which are often executed on the main federated matrix.

On a scenario of training MLogReg on 3 workers and a Critero subset
with 1M rows and ~98M columns after one hot encoding, this patch
reduced the end-to-end execution time by more than 4x but is generally
applicable.

a) STATS BEFORE PATCH:
Total elapsed time:             402.882 sec.
Total compilation time:         2.046 sec.
Total execution time:           400.837 sec.
Cache hits (Mem/Li/WB/FS/HDFS): 4452/0/0/0/0.
Cache writes (Li/WB/FS/HDFS):   4/1576/0/0.
Cache times (ACQr/m, RLS, EXP): 0.085/0.051/0.389/0.000 sec.
HOP DAGs recompiled (PRED, SB): 0/248.
HOP DAGs recompile time:        1.810 sec.
Functions recompiled:           1.
Functions recompile time:       0.253 sec.
Federated I/O (Read, Put, Get): 3/409/134.
Federated Execute (Inst, UDF):  436/6.
Fed Put Count (Sc/Li/Ma/Fr/MC): 0/0/378/0/31.
Fed Put Bytes (Mat/Frame):      1507638888/0 Bytes.
Federated prefetch count:       0.
Total JIT compile time:         29.476 sec.
Total JVM GC count:             18.
Total JVM GC time:              0.442 sec.
Heavy hitter instructions:
  1  m_multiLogReg        370.387      1
  2  fed_r'               221.613     29
  3  fed_mmchain           53.184     63
  4  fed_ba+*              36.619     62
  5  fed_transformencode   17.992      1
  6  n+                    13.264    121
  7  *                     11.882    754
  8  fed_fedinit            8.670      1
  9  -                      4.210    361
 10  +*                     3.892    127

b) STATS AFTER PATCH:
Total elapsed time:             88.907 sec.
Total compilation time:         2.065 sec.
Total execution time:           86.842 sec.
Cache hits (Mem/Li/WB/FS/HDFS): 4510/0/0/0/0.
Cache writes (Li/WB/FS/HDFS):   4/1634/0/0.
Cache times (ACQr/m, RLS, EXP): 0.049/0.034/0.311/0.000 sec.
HOP DAGs recompiled (PRED, SB): 0/248.
HOP DAGs recompile time:        0.506 sec.
Functions recompiled:           1.
Functions recompile time:       0.268 sec.
Federated I/O (Read, Put, Get): 3/380/134.
Federated Execute (Inst, UDF):  378/6.
Fed Put Count (Sc/Li/Ma/Fr/MC): 0/0/378/0/2.
Fed Put Bytes (Mat/Frame):      1507638888/0 Bytes.
Federated prefetch count:       0.
Total JIT compile time:         19.534 sec.
Total JVM GC count:             17.
Total JVM GC time:              0.436 sec.
Heavy hitter instructions:
  1  m_multiLogReg         56.603      1
  2  fed_mmchain           21.669     63
  3  fed_transformencode   18.811      1
  4  fed_ba+*              12.250     62
  5  fed_fedinit            9.061      1
  6  n+                     5.271    121
  7  *                      3.699    754
  8  fed_rightIndex         1.557      2
  9  fed_uack+              1.475      2
 10  +*                     1.330    127

+++ src/main/java/org/apache/sysds/runtime/controlprogram/caching/CacheableData.java
@@ -366,7 +366,7 @@ public abstract class CacheableData<T extends CacheBlock> extends Data
 		return getDataCharacteristics().getCols();
 	}
 	
-	public long getBlocksize() {
+	public int getBlocksize() {
 		return getDataCharacteristics().getBlocksize();
 	}
